to float

+++ nemo/collections/asr/models/label_models.py
@@ -454,7 +454,7 @@ class EncDecSpeakerLabelModel(ModelPT, ExportableEncDecModel):
         device = self.device
         audio = np.array([audio])
         audio_signal, audio_signal_len = (
-            torch.tensor(audio, device=device),
+            torch.tensor(audio, device=device, dtype=torch.float32),
             torch.tensor([audio_length], device=device),
         )
         mode = self.training
